requirement.txt
celery==4.4.5
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.3.0
django-filters==0.2.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
psycopg2==2.8.5
redis==3.5.3
